// themes/GBA/BPGE/EmuDex/src/lib/types.ts
export interface Nature {
  name: string;
  /** Boosted stat index (0=Atk,1=Def,2=Spd,3=SpA,4=SpD). -1 = neutral */
  p: number;
  /** Nerfed stat index. -1 = neutral */
  m: number;
}